tests/wpt/web-platform-tests/trusted-types/window-open.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
<script>
  // helper functions for the tests
  function testWindowOpen(t, win) {
    createURL_policy(window)
        .then(t.step_func_done(p => {
            let url = p.createURL(INPUTS.URL);
            let child_window = win.open(url, "", "");
            child_window.onload = t.step_func_done(_ => {
              assert_equals(child_window.location.href, "" + url);
              child_window.close();
            });
        }));
  }

  async_test(t => {
    testWindowOpen(t, window);
  }, "window.open via policy (successful URL transformation).");

  async_test(t => {
    testWindowOpen(t, document);
  }, "document.open via policy (successful URL transformation).");
</script>
